Reworked memory system - added queueing of messages, custom serialization, changed role of shardMemory implementation in relation to service. Added ability to re-queue tasks.

+++ softwareComponents/taskDistribution/memory/implementation/memoryEntry.hpp
@@ -9,7 +9,7 @@ struct MemoryEntry {
 
     MemoryEntry() {}
 
-    MemoryEntry( unsigned int stamp, int size )
+    MemoryEntry( unsigned int stamp, size_t size )
     : stamp( stamp ){
         stored_data.resize( size );
     }
